rename and fix warnings
