+++ src/tests/decider-distro-tests-rs/tests/utils/test_rpc_server.rs
@@ -1,5 +1,4 @@
-#![feature(async_closure)]
-
+#![allow(dead_code)]
 use eyre::WrapErr;
 use hyper::body::Buf;
 use hyper::service::{make_service_fn, service_fn};
